refactor: decouple the auth manager from the interactive area
It isn't the authorization manager's responsability to play with
modifying the interactive elements. This should be done out of it so
that the authorization manager can be used in other contexts where the
specific UI from the options menu is not present.
